Rename some user attributes to whoami
In all of these, the user being referred is always the current user
and it has to be so, or navigation will lead to parts of the
application which have no useful behavior for the current user.

This is not an exhaustive conversion of these names - only items that
were studied alongside other work.