AE-1815 Add api-key as an attribute of kayttaja
With a corresponding change in etp-core, all kayttaja objects, not
just laatijas may have an api key hash.

api-key is added so that an empty value is set up on initialization
and deserialization so that it gets sent to the backend which expects
it to be there. api-key is not shown in the user interface at this
point.