Fix accessibility issues
The colours in tailwind config aren't the official styleguide colors, as
the site pre-dates the styleguide. Added some of the official colors to
the configuration.

Background color was not set for the whole page. Footer had one gray
shade, while the rest of the page was transparent and thus used the
system default, which incidentally was the same shade as footer at least
for me. Explicitly set the background color for the whole application to
be the official gray shade. It differs slightly from the color used
elsewhere in the site, but not enought to be noticable at least by me.

Links in the footer did not have enough contrast. Added an option to
the Link component to add additional classes, and switched the footer
links to use the darker green meant for small text on non-white
background.

Change history-page in user history had a column for public information.
Public information was in bold green, whilst, non-public information was
in light grey, which didn't have enough contrast on gray rows. Changed the
column to show only the public fields in the same style as all other
information in the table.