AE-1770 Add functions for combined announce
The existing announce stuff went to aria-live.js

Content in annouce.js now provides a more abstract announcement
which the caller can use without explicitly stating what mechanism
it wants. At this time it is assumed that most of the time, it
is best to do both flashMessageStore and an aria-live announcement.