AE-2087 Use gap-x-8 instead of -mx-4/px-4 to implement gutters
The previous implementation was using px-4 in children of a flex row to
have some space between adjacent form fields. To keep the edges of the
forms in place, a negative margin was used in the flex row itself. This
was otherwise pretty OK, but caused some unfortunate interaction with
overflow if the page was narrow and there might not be space for the
-4 margin.

This change removes the previous mechanism and uses just gap-x-8 on the
flex rows to get a similar effect.