Fix docker build failing on fresh clone.
Uberjar file name comes from folder in which was built. Because of this,
the jar used in docker build was one built on host and not in the
builder. This went unnoticed because CI always builds the uberjar before
the docker image, but would fail on fresh clone otherwise.

Fixed by changing directory name in builder image to match repository
naming.

Also added .dockerignore to improve build caching, speed up builds and
prevent such errors in the future.