AE-2087 Use overflow-x-hidden instead of overflow-auto
This seems to prevent wide elements inside the component from
interacting with the size of whole page.

The direct children of the changed element were not separately
scrollable, so it is not entirely clear what the removed overflow-auto
class was doing here. For some reason, the overlow containers deeper
in the DOM will not properly work unless it was there. It is fortunate
that having also overflow-x-hidden works from this point of view.