AE-2132 Grab facus in H1 components
This is quite opinionated in that it expects that any given page
will only have a single H1 element. WCAG shares this opinion, so
we probably want to maintain that constraint in any case.

This should mean that after changing the page in the private site,
tab navigation should always begin from that sole h1 element in the
page. Also the focusing is normally expected to trigger the screen
reader to read the h1 contents aloud.