AE-2132 Fix heading hierarchy for laatija-form
Previously, the form was using multiple h1 headings inside the same
page. WCAG instructs to only have a single h1, and now our H1
component is also beginning to expect a similar invariant.

This change replaces the H1 uses in the form with H2, and adds the
missing H1 into the place where we otherwise have page title.