Fix flakyness in e2e test setup
Docker compose does not care about exit codes of jobs. Migration runner
failed sometimes, perhaps because DB was not ready.

Added a check for migration-runner exit code in e2e start script.

Added a healthcheck for db in compose, and require it to be healthy
before migration runner is run.