AE-2246 Fix aineistoasiakas form
The form would break if the user attempts to use the button for resetting
the API key. The main reason was that it was using the Buffer class which
used to be scope by accident, and which turned out to become invisible
by some unrelated changes.

This commit explicitly depends on and imports the library.

There was also some dodgy syntax on the function, using the `new` keyword
unnecessarily. This was also resolved.