AE-2279: Expose file-key
I have chosen to encapsulate all of the destruction logic in one
namespace. I don't know if it is a good choice as I could also just add
the functionality to destroy anything in their respective namespaces but
as I find those functions to be somewhat dangerous I think encapsulating
the destruction in just a single namespace is safer and therefore I need
to expose this function.