AE-2274: Add capability to destroy viestit
There is a possible problem that is handled in
`check-oikeellisuuden-valvonta-viestiketjut` for now by just logging out
an error if it happens.

There are foreign key reference chain
viestiketju->vo_toimenpide->energiatodistus and
viestiketju->energiatodistus which could possibly cause problems as
viestiketju can reference energiatodistus and a toimenpide that
references different energiatodistus.