AE-2316: Add e2e test for public site to check the visibility of expired energiatodistukset
There might be some problem with the e2e configuration as the sites
layout is somehow broken in the cypress tests but it might also be due
to running the tests with Electron (only thing I can use with WSL). This
is sufficient to test the functionality of the site though.