AE-2490: Show buttons only if the state is reached through initalization
TL;DR; Aborting will only work as one would expect if the signing
process in the backend has failed.

The purpose of being able to call abort is mainly for recovering from
errors and the process should not be typically aborted. The signing
process will continue in the backend if it has not failed even if this
is called. Currenlty this will practically only set the state of the
energiatodistus into draft state. If the signing process in the backend
finishes it will happily set the energiatodistus from draft to signed.