Feature/ae 2490 add confirmation dialog (#713)
* AE-2490: Show instructions only when state is not_started

* AE-2490: Set initial status to correspond the energiatodistus' tila

* AE-2490: Keep track of whether we have reached the current state purely through initalization

* AE-2490: Create the abort function

It is unclear why rightFn is called even if 409 is returned. Maybe this
should be refactored later.

* AE-2490: Show buttons only if the state is reached through initalization

TL;DR; Aborting will only work as one would expect if the signing
process in the backend has failed.

The purpose of being able to call abort is mainly for recovering from
errors and the process should not be typically aborted. The signing
process will continue in the backend if it has not failed even if this
is called. Currenlty this will practically only set the state of the
energiatodistus into draft state. If the signing process in the backend
finishes it will happily set the energiatodistus from draft to signed.

* AE-2490: Remove obsolete CardSigning.svelte

* AE-2490: Remove obsolete prop

* AE-2490: Remove all state specific logic from SigningDialog markup

* AE-2490: Refactor front's signing states

* AE-2490: Create understandable localization strings in Finnish

* AE-2490: Divide logic based on signing state

* AE-2490: Remove unused import

* AE-2490: Add swedish texts

* AE-2490: Fix signing dialog tests

Mostly needed to press the confirmation button

* AE-2490: Fix e2e tests

* AE-2490: fix format

* AE-2490: Update snapshots

* AE-2490: Refine dialog texts

* AE-2490: Refactor signing dialog

SystemSigning.svelte is also redundant and could be refactored fully
into SigningDialog.svelte and signing.js but I don't think it's worth
the effort.