Drop the TODO comment about pdf-route
The version parameter is not actually used for anything. There is
a hidden version dependency in the form that the PDF template file
depends on the version, but that is taken from the ET itself instead
of the request parameters.

This template does not yet exist and thus we won't be able to produce a
PDF, but this is not expected to be a big problem.