AE-2690: Always return the perusparannuspassi
In the case of existing energiatodistus, instead of resurrecting the ppp
we will just always have the ppp if its added even once and just toggle
the validitiy of it. This way we can have the same logic for "add/remove
ppp" in the case of new and existing perusparannuspassi.