AE-2690: Remove `deletePerusparannuspassi`
We will no longer need this functionality in the frontend. Deleting once
a ppp has been added (and thus it exist in the database) is effectively
done by updating the perusparannuspassi so that its valid is set to
false.

We're thinking that the unvalid ones can be deleted when signing.